composer-airflow/requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.3
google-cloud-bigquery
google-cloud-storage
webdriver-manager
selenium==4.9.1